Make tests fast again
Optimize base64url encoding for Node.js performance. Use native Node.js Buffer instead of es-arraybuffer-base64 polyfill when available, as the polyfill is very slow in Node.js.

I checked benchmarks for Storage and it was 8x slower! Just only because polyfill is super slow. Browsers almost all already implements this, so we don't need it, but who knows when Node.js is updated.